requirements.txt
openai
pydantic
python-dotenv
aiohttp
fastapi
uvicorn[standard]
pandas